Move utils to core
