+++ jest.config.js
@@ -15,11 +15,7 @@ module.exports = {
   coveragePathIgnorePatterns: ['/node_modules/', 'enzyme.js'],
   setupFilesAfterEnv: ['<rootDir>/enzyme.js'],
   coverageReporters: ['json', 'lcov', 'text', 'clover'],
-  collectCoverageFrom: [
-    'utils/**/*.js',
-    'views/**/*.js',
-    '!*/**/containers/**'
-  ],
+  collectCoverageFrom: ['views/**/*.js', '!*/**/containers/**'],
   testResultsProcessor: './node_modules/jest-junit-reporter',
   moduleNameMapper: {
     '~/(.*)': '<rootDir>/$1'
