+++ views/blogs/components/post/__tests__/Content.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { shallowWithBaseTheme } from '~/utils/withThemeProviders'
+import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import Content from '~/views/blogs/components/post/Content'
 import theme from '~/theme/blogTheme'
