+++ views/blogs/components/post/postinfo/__tests__/SubInfo.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { shallowWithBaseTheme } from '~/utils/withThemeProviders'
+import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import SubInfo from '../SubInfo'
 import theme from '~/theme/blogTheme'
