+++ views/core/components/__test__/Error.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import Error from '~/views/core/components/Error'
 
 describe('Error', () => {
