+++ views/core/components/__test__/SlideIterator.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import SlideIterator from '~/views/core/components/SlideIterator'
 
 describe('Slide', () => {
