+++ views/core/utils/helpers.js
