+++ views/core/utils/useWindowDimensions.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react'
-import { getWindowDimensions } from '~/utils/helpers'
+import { getWindowDimensions } from '~/views/core/utils/helpers'
 
 /* istanbul ignore next */
 const useWindowDimensions = () => {
