+++ views/core/utils/withThemeProviders.js
