+++ views/home/components/__test__/About.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import About from '~/views/home/components/About'
 
 describe('About', () => {
