+++ views/home/components/__test__/Contact.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import Contact from '~/views/home/components/Contact'
 
 describe('Footer', () => {
