+++ views/home/components/__test__/Home.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import Home from '~/views/home/components/Home'
 import {getTexts} from '~/views/home/components/Home'
 describe('Home', () => {
