+++ views/home/components/__test__/Navbar.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import Navbar from '~/views/home/components/Navbar'
 
 describe('Navbar', () => {
