+++ views/home/components/__test__/Project.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import Project from '~/views/home/components/Project'
 
 describe('Project', () => {
