+++ views/home/components/__test__/WhyMe.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme } from '~/views/core/utils/withThemeProviders'
 import WhyMe from '~/views/home/components/WhyMe'
 
 describe('WhyMe', () => {
