+++ views/home/components/certification/__test__/CertItem.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { shallowWithTheme } from '~/utils/withThemeProviders'
+import { shallowWithTheme } from '~/views/core/utils/withThemeProviders'
 import CertItem from '~/views/home/components/certification/CertItem'
 
 describe('CertItem', () => {
