+++ views/home/components/nav/__test__/NavHeader.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { mountWithTheme, shallowWithTheme } from '~/utils/withThemeProviders'
+import { mountWithTheme, shallowWithTheme } from '~/views/core/utils/withThemeProviders'
 import NavHeader from '~/views/home/components/nav/NavHeader'
 
 describe('Nav Header', () => {
