+++ views/home/components/nav/__test__/NavLang.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { shallowWithTheme } from '~/utils/withThemeProviders'
+import { shallowWithTheme } from '~/views/core/utils/withThemeProviders'
 import NavLang from '~/views/home/components/nav/NavLang'
 
 describe('Nav Lang', () => {
